+++ src/openkit-shared/Protocol/Beacon.cs
@@ -132,10 +132,10 @@ namespace Dynatrace.OpenKit.Protocol
         /// Creates a new instance of type Beacon
         /// </summary>
         /// <param name="logger">Logger for logging messages</param>
-        /// <param name="cache">Cache storing beacon related data</param>
+        /// <param name="beaconCache">Cache storing beacon related data</param>
         /// <param name="configuration">OpenKit related configuration</param>
         /// <param name="clientIPAddress">The client's IP address</param>
-        /// <param name="threadIdProvider">Provider for retrieving thread id</param>
+        /// <param name="threadIDProvider">Provider for retrieving thread id</param>
         /// <param name="timingProvider">Provider for time related methods</param>
         public Beacon(ILogger logger, BeaconCache beaconCache, OpenKitConfiguration configuration, string clientIPAddress,
             IThreadIDProvider threadIDProvider, ITimingProvider timingProvider)
@@ -147,10 +147,10 @@ namespace Dynatrace.OpenKit.Protocol
         /// Creates a new instance of type Beacon
         /// </summary>
         /// <param name="logger">Logger for logging messages</param>
-        /// <param name="cache">Cache storing beacon related data</param>
+        /// <param name="beaconCache">Cache storing beacon related data</param>
         /// <param name="configuration">OpenKit related configuration</param>
         /// <param name="clientIPAddress">The client's IP address</param>
-        /// <param name="threadIdProvider">Provider for retrieving thread id</param>
+        /// <param name="threadIDProvider">Provider for retrieving thread id</param>
         /// <param name="timingProvider">Provider for time related methods</param>
         /// <param name="randomNumberGenerator">Random number generator</param>
         internal Beacon(ILogger logger, BeaconCache beaconCache, OpenKitConfiguration configuration, string clientIPAddress,
@@ -163,13 +163,12 @@ namespace Dynatrace.OpenKit.Protocol
             if (beaconConfiguration.DataCollectionLevel == DataCollectionLevel.USER_BEHAVIOR)
             {
                 SessionNumber = configuration.NextSessionNumber;
-                DeviceID = Truncate(configuration.DeviceID);
+                DeviceID = configuration.DeviceID;
             }
             else
             {
                 SessionNumber = 1;
-                DeviceID = randomNumberGenerator.NextLong(long.MaxValue)
-                    .ToString(CultureInfo.InvariantCulture);
+                DeviceID = randomNumberGenerator.NextLong(long.MaxValue);
             }
 
             this.timingProvider = timingProvider;
@@ -194,7 +193,7 @@ namespace Dynatrace.OpenKit.Protocol
         // *** public properties ***
         public int SessionNumber { get; }
 
-        public string DeviceID { get; }
+        public long DeviceID { get; }
 
         public bool IsEmpty => beaconCache.IsEmpty(SessionNumber);
 
@@ -244,7 +243,7 @@ namespace Dynatrace.OpenKit.Protocol
             {
                 var events = beaconCache.GetEvents(SessionNumber);
                 if (events == null)
-                {                    
+                {
                     return new List<string>();
                 }
 
@@ -266,7 +265,7 @@ namespace Dynatrace.OpenKit.Protocol
                 if (actions == null)
                 {
                     return new List<string>();
-                }                   
+                }
 
                 return actions.Select(x => x.Data).ToList();
             }
@@ -290,7 +289,7 @@ namespace Dynatrace.OpenKit.Protocol
             return $"{WebRequestTagPrefix}_"
                 + $"{ProtocolConstants.ProtocolVersion}_"
                 + $"{httpConfiguration.ServerID}_"
-                + $"{PercentEncoder.Encode(DeviceID, Encoding.UTF8, ReservedCharacters)}_"
+                + $"{DeviceID}_"
                 + $"{SessionNumber}_"
                 + $"{configuration.ApplicationIDPercentEncoded}_"
                 + $"{parentActionID}_"
@@ -750,7 +749,7 @@ namespace Dynatrace.OpenKit.Protocol
             AddKeyValuePairIfValueIsNotNull(builder, BeaconKeyName, TruncateNullSafe(name));
             AddKeyValuePair(builder, BeaconKeyThreadID, threadIDProvider.ThreadID);
         }
-       
+
 
         // helper method for creating basic beacon protocol data
         private string CreateBasicBeaconData()